clean-up

+++ mitoSoft.Holidays.Tests/DouglasAdams/Places.cs
@@ -3,9 +3,9 @@
 namespace mitoSoft.Holidays.Tests.DouglasAdams;
 
 [Flags]
-internal enum Places : ulong
+internal enum Places : byte
 {
     None = 0,
 
-    Earth = 1UL << 0,
+    Earth = 1 << 0,
 }
